refactor(anvil): promote anvil's envelope type to foundry-primitives crate (#12691)
