+++ crates/anvil/Cargo.toml
@@ -28,6 +28,7 @@ foundry-common.workspace = true
 foundry-config.workspace = true
 foundry-evm.workspace = true
 foundry-evm-networks.workspace = true
+foundry-primitives.workspace = true
 
 # alloy
 alloy-evm = { workspace = true, features = ["call-util"] }
