+++ crates/anvil/src/eth/backend/mem/mod.rs
@@ -87,8 +87,7 @@ use anvil_core::eth::{
     block::{Block, BlockInfo},
     transaction::{
         MaybeImpersonatedTransaction, PendingTransaction, ReceiptResponse, TransactionInfo,
-        TypedReceipt, TypedReceiptRpc, TypedTransaction, has_optimism_fields,
-        transaction_request_to_typed,
+        TypedReceipt, TypedReceiptRpc, has_optimism_fields, transaction_request_to_typed,
     },
     wallet::WalletCapabilities,
 };
@@ -108,6 +107,7 @@ use foundry_evm::{
     },
     utils::{get_blob_base_fee_update_fraction, get_blob_base_fee_update_fraction_by_spec_id},
 };
+use foundry_primitives::FoundryTxEnvelope;
 use futures::channel::mpsc::{UnboundedSender, unbounded};
 use op_alloy_consensus::DEPOSIT_TX_TYPE_ID;
 use op_revm::{
@@ -3337,7 +3337,7 @@ impl Backend {
     pub fn get_blob_by_tx_hash(&self, hash: B256) -> Result<Option<Vec<alloy_consensus::Blob>>> {
         // Try to get the mined transaction by hash
         if let Some(tx) = self.mined_transaction_by_hash(hash)
-            && let Ok(typed_tx) = TypedTransaction::try_from(tx)
+            && let Ok(typed_tx) = FoundryTxEnvelope::try_from(tx)
             && let Some(sidecar) = typed_tx.sidecar()
         {
             return Ok(Some(sidecar.sidecar.blobs.clone()));
@@ -3377,7 +3377,7 @@ impl Backend {
         if let Some(full_block) = self.get_full_block(block_id) {
             let sidecar = full_block
                 .into_transactions_iter()
-                .map(TypedTransaction::try_from)
+                .map(FoundryTxEnvelope::try_from)
                 .filter_map(|typed_tx_result| {
                     typed_tx_result.ok()?.sidecar().map(|sidecar| sidecar.sidecar().clone())
                 })
@@ -3645,7 +3645,7 @@ impl TransactionValidator for Backend {
         if let Some(tx_chain_id) = tx.chain_id() {
             let chain_id = self.chain_id();
             if chain_id.to::<u64>() != tx_chain_id {
-                if let TypedTransaction::Legacy(tx) = tx.as_ref() {
+                if let FoundryTxEnvelope::Legacy(tx) = tx.as_ref() {
                     // <https://github.com/ethereum/EIPs/blob/master/EIPS/eip-155.md>
                     if env.evm_env.cfg_env.spec >= SpecId::SPURIOUS_DRAGON
                         && tx.chain_id().is_none()
@@ -3661,7 +3661,7 @@ impl TransactionValidator for Backend {
         }
 
         // Nonce validation
-        let is_deposit_tx = matches!(pending.transaction.as_ref(), TypedTransaction::Deposit(_));
+        let is_deposit_tx = matches!(pending.transaction.as_ref(), FoundryTxEnvelope::Deposit(_));
         let nonce = tx.nonce();
         if nonce < account.nonce && !is_deposit_tx {
             warn!(target: "backend", "[{:?}] nonce too low", tx.hash());
@@ -3672,7 +3672,7 @@ impl TransactionValidator for Backend {
         if env.evm_env.cfg_env.spec >= SpecId::CANCUN && tx.is_eip4844() {
             // Heavy (blob validation) checks
             let blob_tx = match tx.as_ref() {
-                TypedTransaction::EIP4844(tx) => tx.tx(),
+                FoundryTxEnvelope::EIP4844(tx) => tx.tx(),
                 _ => unreachable!(),
             };
 
@@ -3764,7 +3764,7 @@ impl TransactionValidator for Backend {
                 );
             let value = tx.value();
             match tx.as_ref() {
-                TypedTransaction::Deposit(deposit_tx) => {
+                FoundryTxEnvelope::Deposit(deposit_tx) => {
                     // Deposit transactions
                     // https://specs.optimism.io/protocol/deposits.html#execution
                     // 1. no gas cost check required since already have prepaid gas from L1
@@ -3814,7 +3814,7 @@ pub fn transaction_build(
     info: Option<TransactionInfo>,
     base_fee: Option<u64>,
 ) -> AnyRpcTransaction {
-    if let TypedTransaction::Deposit(deposit_tx) = eth_transaction.as_ref() {
+    if let FoundryTxEnvelope::Deposit(deposit_tx) = eth_transaction.as_ref() {
         let dep_tx = deposit_tx;
 
         let ser = serde_json::to_value(dep_tx).expect("could not serialize TxDeposit");
