+++ crates/primitives/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "foundry-primitives"
+
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+alloy-consensus.workspace = true
+alloy-network.workspace = true
+alloy-primitives.workspace = true
+alloy-rlp.workspace = true
+alloy-rpc-types.workspace = true
+alloy-serde.workspace = true
+alloy-evm = { workspace = true, features = ["op"] }
+op-alloy-consensus = { workspace = true, features = ["serde"] }
+op-revm.workspace = true
+revm.workspace = true
+serde_json.workspace = true
