+++ crates/primitives/src/lib.rs
@@ -1,0 +1,3 @@
+//! Foundry primitives
+mod transaction;
+pub use transaction::{FoundryTxEnvelope, FoundryTxType, FoundryTypedTx};
